move operators to own file
